solving the quit

+++ project/project_incremental.py
@@ -7,7 +7,7 @@ import asyncio
 import requests
 import time
 import threading
-import sys
+from numba import jit
 
 class BruteForcer:
     def __init__(self, set_start_method='fork'):
@@ -15,17 +15,11 @@ class BruteForcer:
         self.num_processes = os.cpu_count()
         self.pool = []
         self.result = None
-        
+        self.password = ''
 
-    def queue_control(self, q):
-        while not self.result:
-            if not q.empty():
-                password = q.get()
-                if asyncio.run(self.check_password(password)):
-                    self.result = password
-                    break
-            else:
-                time.sleep(1)   
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        # Çıkış yapıldığında tüm süreçleri sonlandır
+        self.terminate_processes()
 
     def __enter__(self):
         self.get_hashed_password()
@@ -35,17 +29,38 @@ class BruteForcer:
         # queue_control fonksiyonunu ayrı bir thread'de çalıştır
         threading.Thread(target=self.queue_control, args=(q,), daemon=True).start()
 
-        for length in range(3,6):
+        for length in range(2,6):
+            if self.result:
+                break
             self.batch_size = self.calculate_batch_size(self.num_processes, length)
             for i in range(self.num_processes):
                 process = multiprocessing.Process(target=self.brute_force, args=(q, self.batch_size, i, length))
                 process.start()
-                self.pool.append(process)      
+                self.pool.append(process)
             for process in self.pool:
                 process.join()
-
         return self
     
+    def brute_force(self, q, batch_size, offset, length):
+        current = self.calculate_start(batch_size, offset, length)
+        print(f'Process {os.getpid()} started with start: {current} \n')
+        for i in range(batch_size):
+            if self.result:  # Şifre bulunduysa işlemi durdur
+                break
+            hashed_current = hashlib.md5(current.encode()).hexdigest()
+            if self.password == hashed_current:
+                q.put(current)
+            current = self.increment(current)
+
+    def queue_control(self, q):
+        while not self.result:
+            if not q.empty():
+                password = q.get()
+                asyncio.run(self.check_password(password))
+            else:
+                time.sleep(1)
+        self.terminate_processes()
+
     def get_hashed_password(self):
         response = requests.get('http://127.0.0.1:5000/get_password')
         if response.status_code == 200:
@@ -59,30 +74,21 @@ class BruteForcer:
             async with session.post('http://127.0.0.1:5000/check_password', json={'password': password}) as response:
                 if response.status == 200:
                     json_response = await response.json()
-                    return json_response.get('message') == 'Success'
+                    if json_response.get('message') == 'Success':
+                        self.isFound = True
+                        self.result = password
                 else:
                     raise Exception('Failed to check password')
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        # Çıkış yapıldığında tüm süreçleri sonlandır
-        self.terminate_processes()
-
     def terminate_processes(self):
-        # Tüm süreçleri sonlandır
-        for process in self.pool:
-            process.terminate()
-        print("Tüm süreçler sonlandırıldı.")
+            for process in self.pool:
+                if process.is_alive():
+                    process.terminate()
+            print("Tüm süreçler sonlandırıldı.")
 
-    def brute_force(self, q, batch_size, offset, length):
-        current = self.calculate_start(batch_size, offset, length)
-        print(f'Process {os.getpid()} started with start: {current} \n')
-        for i in range(batch_size):
-            hashed_current = hashlib.md5(current.encode()).hexdigest()
-            if self.password == hashed_current:
-                q.put(current)
-            current = self.increment(current)
-
-    def calculate_start(self, batch_size, offset, length):
+    @staticmethod
+    @jit(nopython=True)
+    def calculate_start(batch_size, offset, length):
         def to_base_62(num):
             chars = string.digits + string.ascii_letters
             if num == 0:
@@ -91,25 +97,26 @@ class BruteForcer:
             while num > 0:
                 result.append(chars[num % 62])
                 num = num // 62
-            return "".join(result[::-1])
+            return ''.join(result[::-1])
 
         # Başlangıç numarasını hesapla
         start_num = batch_size * offset
-        
         # Taban 62'ye çevir ve uzunluğa göre sıfırlarla doldur
-        start_num = to_base_62(start_num).rjust(length, string.digits[0])
+        start_num = to_base_62(start_num).zfill(length)
         return start_num
 
     @staticmethod
+    @jit(nopython=True)
     def calculate_batch_size(num_processes, length):
         total_combinations = 62 ** length
         combination_per_process = total_combinations // num_processes
         return combination_per_process
 
     @staticmethod
-    def increment(str):
+    @jit(nopython=True)
+    def increment(base62_str):
         chars = list(string.digits + string.ascii_letters)
-        list_str = list(str)
+        list_str = list(base62_str)
         for i in range(len(list_str)-1, -1, -1):
             if list_str[i] == chars[-1]:
                 list_str[i] = chars[0]
